feat: 목표 상세목록 조회 api 연동 (#142)
* refactor: fetchByGoalId 병렬 fetch -> GoalEditList fetch로 변경 - #138

* refactor: Id 타입 Int64로 변경 - #138

* feat: EditGoalListView 홈에서 보고있는 날짜 이어받기 - #138

* feat: 날짜 변경 시 해당 날짜 goal 불러오는 기능 구현 - #138

* feat: DTO repeatCycle 추가 - #138

* feat: TXCalendar week일 떄 스와이프 콜백 추가 - #138

* feat: TXCalendarUtil dateByAddingWeek 함수 구햔 - #138

* feat: Home, EditGoalList weekCalendar swipe 기능 구현 - #138

---------

+++ Projects/Domain/Goal/Interface/Sources/DTO/GoalPhotoLogListResponseDTO.swift
@@ -9,7 +9,7 @@ import Foundation
 
 /// 목표별 인증샷 목록 응답 DTO입니다.
 public struct GoalPhotoLogListResponseDTO: Decodable {
-    public let goalId: Int
+    public let goalId: Int64
 //    public let myNickname: String
     public let partnerNickname: String
     public let goalTitle: String
@@ -17,7 +17,7 @@ public struct GoalPhotoLogListResponseDTO: Decodable {
 
     public struct PhotoLogResponse: Decodable {
         let photologId: Int
-        let goalId: Int
+        let goalId: Int64
         let imageUrl: String
         let comment: String
         let verificationDate: String
